Move tests/ tree inside ixmp/
