+++ conftest.py
@@ -1,0 +1,20 @@
+import pytest
+
+
+pytest_plugins = ['ixmp.testing']
+
+
+# Hooks
+
+def pytest_addoption(parser):
+    parser.addoption(
+        '--test-r',
+        action='store_true',
+        help='also run tests of the ixmp R interface.',
+    )
+
+
+def pytest_runtest_setup(item):
+    if 'rixmp' in item.keywords and \
+       not item.config.getoption('--test-r'):
+        pytest.skip('skipping rixmp test without --test-r flag')
