+++ ixmp/tests/conftest.py
@@ -18,25 +18,6 @@ from pathlib import Path
 import pytest
 
 
-pytest_plugins = ['ixmp.testing']
-
-
-# Hooks
-
-def pytest_addoption(parser):
-    parser.addoption(
-        '--test-r',
-        action='store_true',
-        help='also run tests of the ixmp R interface.',
-    )
-
-
-def pytest_runtest_setup(item):
-    if 'rixmp' in item.keywords and \
-       not item.config.getoption('--test-r'):
-        pytest.skip('skipping rixmp test without --test-r flag')
-
-
 # Fixtures
 
 @pytest.fixture(scope='session')
@@ -48,4 +29,4 @@ def test_data_path():
 @pytest.fixture(scope='session')
 def tutorial_path():
     """Path to the directory containing the tutorials."""
-    return Path(__file__).parents[1] / 'tutorial'
+    return Path(__file__).parents[2] / 'tutorial'
