+++ ixmp/tests/core/conftest.py
