+++ ixmp/tests/core/test_platform.py
