+++ ixmp/tests/core/test_scenario.py
