+++ ixmp/tests/core/test_timeseries.py
