+++ ixmp/tests/data/test.properties_template
