+++ ixmp/tests/data/test_auth.properties_template
