+++ ixmp/tests/reporting/test_key.py
