+++ ixmp/tests/reporting/test_reporting.py
