+++ ixmp/tests/test_cli.py
