+++ ixmp/tests/test_config.py
