+++ ixmp/tests/test_integration.py
