+++ ixmp/tests/test_model.py
