+++ ixmp/tests/test_tutorials.py
