+++ ixmp/tests/test_utils.py
