+++ setup.cfg
@@ -17,7 +17,6 @@ addopts = --cov=ixmp --cov-config=ci/coveragerc
     -p no:faulthandler
 markers =
     rixmp: test of the ixmp R interface
-testpaths = tests
 
 [aliases]
 test = pytest
